Polyfilled module identifiers are now stable-er.
This affects the value of (require.keys.publicKey.moduleIdentifier)
in polyfilled CommonJS modules.

The module identifiers are now computed relative to a root.
The root is the node_modules directory if one is on the polyfills
path, or else the known root of this project computed relative to the
polyfill.