Use module.id as the path associated with a public key.
This allows node-sec-patterns to reliably double check that a module
exports a key.

Writing github/mikesamuel/attack-review-testbed turned up a
bunch of integration problems in whitelisting modules, and module.id
does provide a reliable basis for optimistic key exchange, and is
what is done by the patched Node runtime used by the testbed.